requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
safetensors>=0.3.0